Modified jmax so that the 'max eclipseprojects' command always updates <project>/.settings/org.eclipse.jdt.core.prefs and <project>/.settings/org.eclipse.jdt.ui.prefs to ensure the same compiler warnings, formatter settings and other miscellaneous Eclipse settings are the same across all projects.

+++ Tele/src/com/sun/max/tele/debug/dump/ELFDumpTeleChannelProtocolAdaptor.java
@@ -104,14 +104,14 @@ public class ELFDumpTeleChannelProtocolAdaptor extends TeleChannelDataIOProtocol
                 int namesz = dis.read_Elf64_Word();
                 final int descsz = dis.read_Elf64_Word();
                 final int type = dis.read_Elf64_Word();
-                final String name = readNoteString(dis, (int) namesz);
+                final String name = readNoteString(dis, namesz);
                 readLength += 12 + namesz;
                 while (namesz % 8 != 0) {
                     dis.read_Elf32_byte();
                     readLength++;
                     namesz++;
                 }
-                final byte[] desc = readNoteDesc(dis, (int) descsz);
+                final byte[] desc = readNoteDesc(dis, descsz);
                 readLength += descsz;
 
                 entryHandler.processNoteEntry(type, name, desc);
